1-requirements.txt
azure-mgmt-resource
azure-mgmt-compute
azure-mgmt-network
azure-identity
azure-storage-blob
azure-mgmt-monitor